requirements.txt
gradio==4.18.0
opencv-python
tensorflow
keras
numpy
Pillow
